Add try/finally to python dist integration testing.

+++ tests/python/pants_test/backend/python/tasks/test_python_distribution_integration.py
@@ -20,17 +20,19 @@ class PythonDistributionIntegrationTest(PantsRunIntegrationTest):
   fasthello_install_requires = 'testprojects/src/python/python_distribution/fasthello_with_install_requires'
 
   def test_pants_binary(self):
-    command=['binary', '{}:main'.format(self.fasthello_project)]
-    pants_run = self.run_pants(command=command)
-    self.assert_success(pants_run)
-    # Check that the pex was built.
     pex = os.path.join(get_buildroot(), 'dist', 'main.pex')
-    self.assertTrue(os.path.isfile(pex))
-    # Check that the pex runs.
-    output = subprocess.check_output(pex)
-    self.assertIn('Super hello', output)
-    # Cleanup
-    os.remove(pex)
+    try:
+      command=['binary', '{}:main'.format(self.fasthello_project)]
+      pants_run = self.run_pants(command=command)
+      self.assert_success(pants_run)
+      # Check that the pex was built.
+      self.assertTrue(os.path.isfile(pex))
+      # Check that the pex runs.
+      output = subprocess.check_output(pex)
+      self.assertIn('Super hello', output)
+    finally:
+      if os.path.exists(pex):
+        os.remove(pex)
 
   def test_pants_run(self):
     command=['run', '{}:main'.format(self.fasthello_project)]
@@ -45,17 +47,20 @@ class PythonDistributionIntegrationTest(PantsRunIntegrationTest):
     self.assert_success(pants_run)
 
   def test_with_install_requires(self):
-    command=['run', '{}:main_with_no_conflict'.format(self.fasthello_install_requires)]
-    pants_run = self.run_pants(command=command)
-    self.assert_success(pants_run)
-    self.assertIn('United States', pants_run.stdout_data)
-    command=['binary', '{}:main_with_no_conflict'.format(self.fasthello_install_requires)]
-    pants_run = self.run_pants(command=command)
-    self.assert_success(pants_run)
     pex = os.path.join(get_buildroot(), 'dist', 'main_with_no_conflict.pex')
-    output = subprocess.check_output(pex)
-    self.assertIn('United States', output)
-    os.remove(pex)
+    try:
+      command=['run', '{}:main_with_no_conflict'.format(self.fasthello_install_requires)]
+      pants_run = self.run_pants(command=command)
+      self.assert_success(pants_run)
+      self.assertIn('United States', pants_run.stdout_data)
+      command=['binary', '{}:main_with_no_conflict'.format(self.fasthello_install_requires)]
+      pants_run = self.run_pants(command=command)
+      self.assert_success(pants_run)
+      output = subprocess.check_output(pex)
+      self.assertIn('United States', output)
+    finally:
+      if os.path.exists(pex):
+        os.remove(pex)
 
   def test_with_conflicting_transitive_deps(self):
     command=['run', '{}:main_with_conflicting_dep'.format(self.fasthello_install_requires)]
@@ -70,45 +75,51 @@ class PythonDistributionIntegrationTest(PantsRunIntegrationTest):
     self.assertIn('fasthello', pants_run.stderr_data)
 
   def test_pants_binary_dep_isolation_with_multiple_targets(self):
-    command=['binary', '{}:main_with_no_conflict'.format(self.fasthello_install_requires),
-             '{}:main_with_no_pycountry'.format(self.fasthello_install_requires)]
-    pants_run = self.run_pants(command=command)
-    self.assert_success(pants_run)
-    # Check that the pex was built.
     pex1 = os.path.join(get_buildroot(), 'dist', 'main_with_no_conflict.pex')
-    self.assertTrue(os.path.isfile(pex1))
     pex2 = os.path.join(get_buildroot(), 'dist', 'main_with_no_pycountry.pex')
-    self.assertTrue(os.path.isfile(pex2))
-    # Check that the pex 1 runs.
-    output = subprocess.check_output(pex1)
-    self.assertIn('Super hello', output)
-    # Check that the pex 2 fails due to no python_dists leaked into it.
     try:
-      output = subprocess.check_output(pex2)
-    except subprocess.CalledProcessError as e:
-      self.assertNotEquals(0, e.returncode)
-    # Cleanup
-    os.remove(pex1)
-    os.remove(pex2)
+      command=['binary', '{}:main_with_no_conflict'.format(self.fasthello_install_requires),
+               '{}:main_with_no_pycountry'.format(self.fasthello_install_requires)]
+      pants_run = self.run_pants(command=command)
+      self.assert_success(pants_run)
+      # Check that the pex was built.
+      self.assertTrue(os.path.isfile(pex1))
+      self.assertTrue(os.path.isfile(pex2))
+      # Check that the pex 1 runs.
+      output = subprocess.check_output(pex1)
+      self.assertIn('Super hello', output)
+      # Check that the pex 2 fails due to no python_dists leaked into it.
+      try:
+        output = subprocess.check_output(pex2)
+      except subprocess.CalledProcessError as e:
+        self.assertNotEquals(0, e.returncode)
+    finally:
+      # Cleanup
+      if os.path.exists(pex1):
+        os.remove(pex1)
+      if os.path.exists(pex2):
+        os.remove(pex2)
 
   def test_pants_resolves_local_dists_for_current_platform_only(self):
     # Test that pants will override pants.ini platforms config when building
     # or running a target that depends on native (c or cpp) sources.
-    pants_ini_config = {'python-setup': {'platforms': ['current', 'linux-x86_64']}}
-
-    # Clean all to rebuild requirements pex.
-    command=['clean-all', 'run', '{}:main'.format(self.fasthello_project)]
-    pants_run = self.run_pants(command=command, config=pants_ini_config)
-    self.assert_success(pants_run)
-
-    command=['binary', '{}:main'.format(self.fasthello_project)]
-    pants_run = self.run_pants(command=command, config=pants_ini_config)
-    self.assert_success(pants_run)
-    # Check that the pex was built.
     pex = os.path.join(get_buildroot(), 'dist', 'main.pex')
-    self.assertTrue(os.path.isfile(pex))
-    # Check that the pex runs.
-    output = subprocess.check_output(pex)
-    self.assertIn('Super hello', output)
-    # Cleanup
-    os.remove(pex)
+    pants_ini_config = {'python-setup': {'platforms': ['current', 'linux-x86_64']}}
+    try:
+      # Clean all to rebuild requirements pex.
+      command=['clean-all', 'run', '{}:main'.format(self.fasthello_project)]
+      pants_run = self.run_pants(command=command, config=pants_ini_config)
+      self.assert_success(pants_run)
+
+      command=['binary', '{}:main'.format(self.fasthello_project)]
+      pants_run = self.run_pants(command=command, config=pants_ini_config)
+      self.assert_success(pants_run)
+      # Check that the pex was built.
+      self.assertTrue(os.path.isfile(pex))
+      # Check that the pex runs.
+      output = subprocess.check_output(pex)
+      self.assertIn('Super hello', output)
+    finally:
+      if os.path.exists(pex):
+        # Cleanup
+        os.remove(pex)
